Try out hybrid Jackson/byte based sorting.
Not great, but the Genome object isn't a good example with the current
implementation. Either soley a byte sort, each member of the features
array is sorted one by one in Jackon, or the whole thing is sorted by
Jackson after a pass through.

With array optimization nad other heuristics it could be useful though.