KBASE-922 #comment add X-Forwarded-For & X-Real-IP handling to Java
By default, the logged ip address will be, in order of preference:
1) X-Forwarded-For[0]
2) X-Real-IP
3) The standard address

If the config file for the server includes 
dont_trust_x_ip_headers = true
then only the standard address is logged.

The contents of X-Forwarded-For is always logged, if it exists, in a log
message preceting the start method message.

Also did some warning cleanup.