Make several servlet methods public for reuse
Several of the JsonServerServlet methods are generally useful for other
services running in the same WAR file. Provide a public interface for
these methods and for necessary constants.

Namely:
KB_DEP
getConfig
getIpAddress
setUpResponseHeaders

Was considering logHeaders, but it's too simple. If it had taken the
request object as an argument, maybe.

TODO: testing of the exposed methods.