Workarounds for Shock startup & admin
Adds workarounds on a per version basis get to get Shock to start up
without human intervention (some versions of Shock ask if schemas should
be upgraded if the DB is empty and waits for a response).

Also inserts an admin into the Users database since some versions of
Shock won't add an admin until 1) the user has used the API and
therefore has a record in the DB and 2) Shock is started with the name
of that user in the config file. Hence there cannot be an admin the
first time Shock starts.

Version detection is current done by checking the md5 of the
shock-server binary and checking a hardcoded map to get the version,
which is an utterly crap system, but I don't have a better idea at the
moment.