Add constructor for embedded JsonServerServlet
The only constructor for the JsonServerServlet assumes it's running on
its own - is parses a config file, sets up its own auth, etc.

Add a constructor where the server build is more controllable and
therefore the server is more embeddable.

Allowing the logger to be specified might also be useful, but so much of
the logger api is used that might be tricky.

Also add some happy path tests.