Fix bugs in the JSON tokenizer used by JAVA SDK apps/services
SDK based apps, services, and clients use a customized JSON tokenizer
from this repo (we'll call it JTS). The tokenizer wraps an inner Jackson
tokenizer (inner). The JTS keeps track of the current path in the JSON
in its own state.

The bugs result when one of several helper methods (the next*() methods)
are called on JTS. These methods delegate to the inner tokenizer and
therefore the state stored in JTS is not updated even though the token
stream provided by inner is advanced. This causes lots of weird errors
and can make the tokenizer unusable in the context of SDK processes.

The fix is to make the helper methods delegate to the JTS superclass
instead of the inner tokenizer. The superclass methods all call
nextToken() under the hood, which is overridden in JTS, and ensures the
state is updated correctly.

In one case, the super class method was abstract, so the implementation
was copied from MinimalParserBase, a subclass of the superclass.

I assume that no code paths actually used the helper methods as we
haven't seen these failures in the field, as far as I know. However, I
did start seeing failures caused by these methods after updating Jackson
to 2.9.9, which is why I'm fixing these issues.